updated create_objective_func to take a map_func arg
The optional map_func is a mechanism for client code to inspect and/or map theta onto new values, prior to prior_func being called. It's a way of potentially supporting the fitting of mass/age params to SED through mapping to teff/logg from MIST models.

Also with this change prior_func is expected to return np.inf on "failure" and a finite value on success (which is added to the return value of sim_func to make the objective_func return value)