_fit() funcs take full theta0 & fit_mask to indicate what is to be fitted
Makes things easier to understand and there's only one "type" of theta; the full set of params to generate a model SED.

Tested out this approach with fitting of various combinations of params (in fit_mask). This highlighted bugs in the previous approach which used np.putmask() to coallesce theta for each step (this appears to not like non-contiguous bitmasks).