simplified the application of filter transmission
We had a strange mix of normalized transmission (values between 0 and 1) which was then adjusted to total 1 prior to being applied to the fluxes. The norrmalization step isn't needed, so I've just modified the Norm-Transmission field appended to each filter table so the transmission values total to 1 (normalized to AUC of 1).